Arreglos varios

+++ pf/src/entidad/Util.java
@@ -191,7 +191,7 @@ public class Util {
     
     public static boolean isNumeric(String cadena){
     	try {
-    		Integer.parseInt(cadena);
+    		Long.parseLong(cadena);
     		return true;
     	} catch (NumberFormatException nfe){
     		return false;
